requirements.txt
chino==2.13.0
python-dotenv==1.0.0
pytest==7.4.0
